+++ src/api/users/controllers/post.js
@@ -1,17 +1,10 @@
-const express = require("express");
-const user = require("src/models/user.js")
+import express from "express";
+const app = express();
+import userRouters from "../routers/index.js";
 
-const router = express.Router();
-
-//crear usuario
-router.post(, (req, res) => {})
-
-app.post("./api/post.js", (request, response, error) => {
-  response.send("Status:200");
-  response.json("Ok");
-  request.body("Ok")
-});
+//middleware
+app.use("/api", userRouters);
 
 
-
+// revisar archivo en mentoria Si lo dejo aca no me funciona
