--- src/api/users/routes/index.js
@@ -1,9 +1,0 @@
-// todas las rutas de la entidad 
-const express = require ("express");
-
-const User = require(src\models\user.js)
-
-const router = express.Router();
-
-//Creación de un usuario
-router.post()
